Refactor app into modules
